errors in MapsActivity that i need help with

+++ app/src/main/java/com/example/project1_daytripplanner_montgomery/places.kt
@@ -7,6 +7,6 @@ data class places (
     val address : String,
     val pricePoint : String,
     val url : String,
-    val lat : Float,
-    val long : Float
+    val lat : Double,
+    val long : Double
 )
